sdks/python/playwright/requirements.txt
eyes-playwright>=5.15.1
pytest-playwright>=0.3.1